modificado tipos de datos de funciones de operaciones

+++ TP1/funciones.h
@@ -2,24 +2,24 @@
 #define FUNCIONES_H_INCLUDED
 
 //imprime el menu de la calculadora con los 2 numeros ingresados actualizados
-void imprimirMenu(int primerNumero,int segundoNumero);
+void imprimirMenu(float primerNumero,float segundoNumero);
 
 //Devuelve la suma de 2 numeros
-int suma(int primerNumero, int segundoNumero);
+float suma(float primerNumero,float segundoNumero);
 
 //Devuelve la resta de 2 numeros
-int resta(int primerNumero, int segundoNumero);
+float resta(float primerNumero,float segundoNumero);
 
 //Devuelve la multiplicacion de 2 numeros
-int multiplicar(int primerNumero, int segundoNumero);
+float multiplicar(float primerNumero,float segundoNumero);
 
 //Devuelve la division de 2 numeros
-float dividir(int primerNumero, int segundoNumero);
+float dividir(float primerNumero,float segundoNumero);
 
 //devuelve el factorial del primer numero ingresado
 long int calcularFactorial(int num);
 
 //realizar la operacion que elije el usuario con los 2 numeros ingresados
-void realizarOperacion(int opcion,int primerNumero,int segundoNumero);
+void realizarOperacion(int opcion,float primerNumero,float segundoNumero);
 
 #endif // FUNCIONES_H_INCLUDED
